requirements.txt

google-generativeai==0.4.2
python-dotenv>=1.0.1
typer[all]>=0.12.3
